+++ tests/integration/rpc/backend/test_call_tx.go
@@ -12,6 +12,7 @@ import (
 	"github.com/ethereum/go-ethereum/rlp"
 	"google.golang.org/grpc/metadata"
 
+	rpcbackend "github.com/cosmos/evm/rpc/backend"
 	"github.com/cosmos/evm/rpc/backend/mocks"
 	rpctypes "github.com/cosmos/evm/rpc/types"
 	"github.com/cosmos/evm/testutil/constants"
@@ -381,7 +382,7 @@ func (s *TestSuite) TestSendRawTransaction() {
 			name: "fail - no EVM mempool",
 			registerMock: func() {
 				s.backend.AllowUnprotectedTxs = true
-				s.backend.Mempool = nil
+				s.backend.Mempool = rpcbackend.NoOpMempool{}
 			},
 			rawTx:    func() []byte { return rlpEncodedBz },
 			expHash:  common.Hash{},